# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AegisAl Knowledge Security Assistant",
    "description": "A source-available LangGraph knowledge security assistant by ShenSiQi.",
    "version": "1.3.4"
  },
  "paths": {
...
    "/api/assistant/artifacts/{artifact_id}": {
      "get": {
        "summary": "Download Assistant Artifact",
        "operationId": "download_assistant_artifact_api_assistant_artifacts__artifact_id__get",
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 120,
              "default": "default",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/assistant/questions": {
      "post": {
        "summary": "Ask",
        "operationId": "ask_api_assistant_questions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assistant/questions/stream": {
      "post": {
        "summary": "Ask Stream",
        "operationId": "ask_stream_api_assistant_questions_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/assistant/interrupts/resume": {
      "post": {
        "summary": "Resume Assistant Interrupt",
        "operationId": "resume_assistant_interrupt_api_assistant_interrupts_resume_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantInterruptResumeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/assistant/conversations": {
      "get": {
        "summary": "List Assistant Conversations",
        "operationId": "list_assistant_conversations_api_assistant_conversations_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 120,
              "default": "default",
              "title": "User Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          },
          {
            "name": "archived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Archived"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assistant/conversations/{session_id}": {
      "get": {
        "summary": "Assistant Conversation Detail",
        "operationId": "assistant_conversation_detail_api_assistant_conversations__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 120,
              "default": "default",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Assistant Conversation",
        "operationId": "delete_assistant_conversation_api_assistant_conversations__session_id__delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 120,
              "default": "default",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assistant/conversations/{session_id}/exchanges/{exchange_id}/table-edits": {
      "patch": {
        "summary": "Update Assistant Conversation Table Edits",
        "operationId": "update_assistant_conversation_table_edits_api_assistant_conversations__session_id__exchanges__exchange_id__table_edits_patch",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "exchange_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Exchange Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 120,
              "default": "default",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantStructuredDataEditRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assistant/conversations/{session_id}/archive": {
      "post": {
        "summary": "Archive Assistant Conversation",
        "operationId": "archive_assistant_conversation_api_assistant_conversations__session_id__archive_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 120,
              "default": "default",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantConversationArchiveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assistant/short-term-sessions/{session_id}": {
      "delete": {
        "summary": "Clear Assistant Short Term Session",
        "operationId": "clear_assistant_short_term_session_api_assistant_short_term_sessions__session_id__delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 120,
              "default": "default",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "success"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "data": {
            "title": "Data"
          }
        },
        "type": "object",
        "title": "ApiResponse"
      },
...
      "AskAttachment": {
        "properties": {
          "file_name": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "File Name"
          },
          "content": {
            "type": "string",
            "maxLength": 120000,
            "minLength": 1,
            "title": "Content"
          },
          "mime_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Mime Type"
          }
        },
        "type": "object",
        "required": [
          "file_name",
          "content"
        ],
        "title": "AskAttachment"
      },
      "AskRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question"
          },
          "top_k": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 5
          },
          "user_id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "User Id",
            "default": "default"
          },
          "session_id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Session Id",
            "default": "default"
          },
          "response_language": {
            "type": "string",
            "maxLength": 24,
            "title": "Response Language",
            "default": "zh-Hans"
          },
          "emoji_mode": {
            "type": "string",
            "enum": [
              "off",
              "moderate",
              "active"
            ],
            "title": "Emoji Mode",
            "default": "moderate"
          },
          "intent_hint": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "component_vulnerability_catalog",
                  "recent_high_vulnerability_lookup",
                  "information_consultation"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Intent Hint"
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/AskAttachment"
            },
            "type": "array",
            "maxItems": 300,
            "title": "Attachments"
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "AskRequest"
      },
      "AssistantConversationArchiveRequest": {
        "properties": {
          "archived": {
            "type": "boolean",
            "title": "Archived"
          }
        },
        "type": "object",
        "required": [
          "archived"
        ],
        "title": "AssistantConversationArchiveRequest"
      },
      "AssistantDataTableEdit": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Id"
          },
          "type": {
            "type": "string",
            "maxLength": 80,
            "title": "Type",
            "default": "records-table"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          },
          "columns": {
            "items": {
              "$ref": "#/components/schemas/AssistantDataTableEditColumn"
            },
            "type": "array",
            "maxItems": 64,
            "minItems": 1,
            "title": "Columns"
          },
          "rows": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "maxItems": 200,
            "title": "Rows"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          },
          "edited": {
            "type": "boolean",
            "title": "Edited",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "columns"
        ],
        "title": "AssistantDataTableEdit"
      },
      "AssistantDataTableEditColumn": {
        "properties": {
          "key": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Key"
          },
          "label": {
            "type": "string",
            "maxLength": 240,
            "minLength": 1,
            "title": "Label"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "date",
                  "link",
                  "number",
                  "status",
                  "tags",
                  "text"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "editable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Editable"
          }
        },
        "type": "object",
        "required": [
          "key",
          "label"
        ],
        "title": "AssistantDataTableEditColumn"
      },
      "AssistantInterruptResumeRequest": {
        "properties": {
          "thread_id": {
            "type": "string",
            "maxLength": 160,
            "minLength": 8,
            "title": "Thread Id"
          },
          "interrupt_id": {
            "type": "string",
            "maxLength": 160,
            "title": "Interrupt Id",
            "default": ""
          },
          "decision": {
            "type": "string",
            "enum": [
              "confirm",
              "cancel"
            ],
            "title": "Decision"
          },
          "format": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "md",
                  "html",
                  "docx",
                  "xlsx",
                  "pdf",
                  "all"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Format"
          },
          "user_id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "User Id",
            "default": "default"
          },
          "session_id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Session Id",
            "default": "default"
          },
          "response_language": {
            "type": "string",
            "maxLength": 24,
            "title": "Response Language",
            "default": "zh-Hans"
          }
        },
        "type": "object",
        "required": [
          "thread_id",
          "decision"
        ],
        "title": "AssistantInterruptResumeRequest",
        "description": "Resume any assistant-owned LangGraph interrupt without coupling the client to one subgraph."
      },
      "AssistantStructuredDataEditRequest": {
        "properties": {
          "tables": {
            "items": {
              "$ref": "#/components/schemas/AssistantDataTableEdit"
            },
            "type": "array",
            "maxItems": 12,
            "minItems": 1,
            "title": "Tables"
          }
        },
        "type": "object",
        "required": [
          "tables"
        ],
        "title": "AssistantStructuredDataEditRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}